+++ lib/src/core/definitions/type_definitions.dart
@@ -47,10 +47,12 @@ typedef QuestionnaireBooleanWidgetBuilder = Widget Function(
 /// [index]: The index of the item in the questionnaire.
 /// [isLastItem]: Determines if this is the last item in the questionnaire.
 /// [questionnaireItem]: The questionnaire group item to be displayed.
+/// [childrenAssigner]: A function to assign deafult widget to the group item children.
 typedef QuestionnaireGroupWidgetBuilder = Widget Function(
   int index,
   bool isLastItem,
   QuestionnaireItem questionnaireItem,
+  Widget Function(QuestionnaireItem questionnaireItem) childrenAssigner,
 );
 
 /// A typedef for building a choice widget for a questionnaire.