+++ lib/src/ui/components/slivers/questionnaire_sliver_item_wrapper.dart
@@ -105,6 +105,13 @@ final class QuestionnaireSliverItemWrapper extends QuestionnaireItemWrapper {
             index,
             isLastItem,
             questionnaireItem,
+            (questionnaireItem) {
+              return QuestionnaireItemWrapper(
+                index: index,
+                questionnaireItem: questionnaireItem,
+                isLastItem: isLastItem,
+              );
+            },
           );
         }
         return QuestionnaireSliverGroupItem(